# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OptiPlan360 API",
    "description": "Comprehensive API documentation for OptiPlan360 - Integrated Woodworking Management System",
    "version": "1.0.0",
    "contact": {
      "name": "OptiPlan360 Development Team",
      "email": "dev@optiplan360.com",
      "url": "https://optiplan360.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.optiplan360.com/v1",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.optiplan360.com/v1",
      "description": "Staging Server"
    }
  ],
  "paths": {
...
    "/ocr/upload": {
      "post": {
        "tags": ["OCR"],
        "summary": "Upload Document for OCR Processing",
        "description": "Upload a document and process it with OCR",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document file to process"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OCR processing initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OCRResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "example": "ocr_task_123456"
          },
          "status": {
            "type": "string",
            "enum": ["processing", "completed", "failed", "requires_review"],
            "example": "processing"
          },
          "confidence_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 85.5
          },
          "extracted_data": {
            "type": "object",
            "example": {
              "customer_name": "Test Customer",
              "product_code": "OAK001",
              "quantity": 100
            }
          }
        }
      },
...
    }
  },
...
}